server/requirements.txt
elasticsearch
flask~=2.2.2
flask-jwt-extended
cassandra-driver
torch
torchvision
numpy
pillow